Rename Data.Serialize to Data.MutByteArray
