+++ core/src/Streamly/Internal/Data/Serialize/TH/RecHeader.hs
@@ -27,7 +27,7 @@ import Data.Maybe (fromJust)
 import Language.Haskell.TH
 import Streamly.Internal.Data.Serialize.Type (Serialize(..))
 import Data.Foldable (foldlM)
-import Streamly.Internal.Data.MutByteArray (MutByteArray)
+import Streamly.Internal.Data.MutByteArray.Type (MutByteArray)
 import Data.Proxy (Proxy(..))
 
 import qualified Streamly.Internal.Data.Unbox as Unbox
