+++ core/streamly-core.cabal
@@ -292,7 +292,7 @@ library
                      , Streamly.Internal.Data.Producer
 
                      -- streamly-core-array-types
-                     , Streamly.Internal.Data.Serialize
+                     , Streamly.Internal.Data.MutByteArray
                      , Streamly.Internal.Data.Serialize.Parser
                      , Streamly.Internal.Data.Serialize.Stream
 
@@ -367,7 +367,7 @@ library
                      -- streamly-core released modules in alphabetic order
                      -- NOTE: these must be added to streamly.cabal as well
                      , Streamly.Console.Stdio
-                     , Streamly.Data.Serialize
+                     , Streamly.Data.MutByteArray
                      , Streamly.Data.Array
                      , Streamly.Data.Array.Generic
                      , Streamly.Data.MutArray
@@ -421,9 +421,10 @@ library
                     , Streamly.Internal.Data.Array.Type
 
                     -- Unboxed IORef
-                    , Streamly.Internal.Data.MutByteArray
-                    , Streamly.Internal.Data.Unbox
                     , Streamly.Internal.Data.IORef.Unboxed
+
+                    , Streamly.Internal.Data.MutByteArray.Type
+                    , Streamly.Internal.Data.Unbox
                     , Streamly.Internal.Data.Unbox.TH
                     , Streamly.Internal.Data.Serialize.Type
                     , Streamly.Internal.Data.Serialize.TH
