+++ src/Streamly/Internal/Data/Stream/IsStream/Common.hs
@@ -79,7 +79,7 @@ import Streamly.Internal.Data.Stream.IsStream.Type
 import Streamly.Internal.Data.Stream.Serial (SerialT)
 import Streamly.Internal.Data.Time.Units (AbsTime, RelTime64, addToAbsTime64)
 import Streamly.Internal.System.IO (defaultChunkSize)
-import Streamly.Data.Serialize (Unbox)
+import Streamly.Data.MutByteArray (Unbox)
 
 import qualified Streamly.Internal.Data.Array as A
 import qualified Streamly.Internal.Data.Stream.Async as Async
