+++ src/Streamly/Internal/Network/Inet/TCP.hs
@@ -111,7 +111,7 @@ import Streamly.Internal.Data.Array (pinnedWriteNUnsafe)
 import Streamly.Internal.Data.Fold ( Fold(..) )
 import Streamly.Data.Stream (Stream)
 import Streamly.Internal.Data.Tuple.Strict (Tuple'(..))
-import Streamly.Data.Serialize (Unbox)
+import Streamly.Data.MutByteArray (Unbox)
 import Streamly.Data.Unfold (Unfold)
 import Streamly.Internal.Network.Socket (SockSpec(..), accept, acceptor)
 import Streamly.Internal.System.IO (defaultChunkSize)
