+++ test/Streamly/Test/Data/Serialize/CompatV0.hs
@@ -10,8 +10,8 @@ module Streamly.Test.Data.Serialize.CompatV0
 --------------------------------------------------------------------------------
 
 import Test.QuickCheck (Arbitrary, arbitrary, elements)
-import Streamly.Internal.Data.Serialize (Serialize)
-import qualified Streamly.Internal.Data.Serialize as Serialize
+import Streamly.Internal.Data.MutByteArray (Serialize)
+import qualified Streamly.Internal.Data.MutByteArray as Serialize
 
 --------------------------------------------------------------------------------
 -- Types
